requirements.txt
# Core
openai>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# UI
streamlit>=1.35.0

# Testing
pytest>=8.0.0
